separate the work
