+++ packages/zowe-explorer/i18n/sample/src/utils/ProfileManagement.i18n.json
@@ -1,0 +1,16 @@
+{
+  "profiles.operation.cancelled": "Operation Cancelled",
+  "editProfileQpItem.editProfile.qpLabel": "$(pencil) Edit profile",
+  "editProfileQpItem.editProfile.qpDetail": "Update profile connection information",
+  "updateBasicAuthQpItem.updateCredentials.qpLabel": "$(refresh) Update Credentials",
+  "updateBasicAuthQpItem.updateCredentials.qpDetail": "Update stored username and password",
+  "addBasicAuthQpItem.addCredentials.qpLabel": "$(plus) Add Credentials",
+  "addBasicAuthQpItem.addCredentials.qpDetail": "Add username and password for basic authentication",
+  "loginQpItem.login.qpLabel": "$(arrow-right) Log in to authentication service",
+  "loginQpItem.login.qpDetail": "Log in to obtain a new token value",
+  "logoutQpItem.logout.qpLabel": "$(arrow-left) Log out of authentication service",
+  "logoutQpItem.logout.qpDetail": "Log out to invalidate and remove stored token value",
+  "qpPlaceholders.qp.basic": "Profile is using basic authentication. Choose a profile action.",
+  "qpPlaceholders.qp.token": "Profile is using token authentication. Choose a profile action.",
+  "qpPlaceholders.qp.choose": "Profile doesn't specify an authentication method. Choose a profile action."
+}
