+++ packages/zowe-explorer/i18n/sample/src/utils/ProfilesUtils.i18n.json
@@ -21,19 +21,5 @@
   "initializeZoweFolder.error": "Failed to initialize Zowe folder: {0}",
   "initializeZoweProfiles.success": "Zowe Profiles initialized successfully.",
   "initializeZoweTempFolder.success": "Zowe Temp folder initialized successfully.",
-  "profiles.operation.cancelled": "Operation Cancelled",
-  "editProfileQpItem.editProfile.qpLabel": "$(pencil) Edit profile",
-  "editProfileQpItem.editProfile.qpDetail": "Update profile connection information",
-  "updateBasicAuthQpItem.updateCredentials.qpLabel": "$(refresh) Update Credentials",
-  "updateBasicAuthQpItem.updateCredentials.qpDetail": "Update stored username and password",
-  "addBasicAuthQpItem.addCredentials.qpLabel": "$(plus) Add Credentials",
-  "addBasicAuthQpItem.addCredentials.qpDetail": "Add username and password for basic authentication",
-  "loginQpItem.login.qpLabel": "$(arrow-right) Login to authentication service",
-  "loginQpItem.login.qpDetail": "Login to obtain and update stored token value",
-  "logoutQpItem.logout.qpLabel": "$(arrow-left) Log out of authentication service",
-  "logoutQpItem.logout.qpDetail": "Log out to invalidate and remove stored token value",
-  "qpPlaceholders.qp.basic": "Profile is using basic authentication. Choose a profile action.",
-  "qpPlaceholders.qp.token": "Profile is using token authentication. Choose a profile action.",
-  "qpPlaceholders.qp.choose": "Profile doesn't specify an authentication method. Choose a profile action.",
   "getProfile.notTreeItem": "Tree Item is not a Zowe Explorer item."
 }
